python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
azure-identity==1.24.0
azure-ai-projects==1.0.0
langchain==0.3.27
langchain-openai==0.3.30
langgraph==0.6.5
langchain-core==0.3.74
aiosqlite==0.19.0
python-multipart==0.0.6
